moved zip function to new file
