+++ src/components/DownloadDataButton.js
@@ -1,11 +1,8 @@
 // Download button for individual plots
-import { saveAs } from "file-saver";
 import React from "react";
 import { Button } from "react-bootstrap";
 import "./DownloadButton.css";
-
-import JSZip from "jszip";
-import JSZipUtils from "jszip-utils";
+import downloadzip from "./ZipFiles";
 
 export default function DownloadDataButton({
   plotvars,
@@ -14,32 +11,6 @@ export default function DownloadDataButton({
   plottypes,
   regions,
 }) {
-  // zip multiple images
-  const downloadzip = (datalist) => {
-    var zip = new JSZip();
-    var count = 0;
-    var zipFilename = "UKCORDEX-data.zip";
-    var urls = datalist;
-
-    urls.map((url) => {
-      // get the right basename for each image
-      var filename = url.slice(url.lastIndexOf("/") + 1);
-      // loading a file and add it in a zip file
-      JSZipUtils.getBinaryContent(url, function (err, data) {
-        if (err) {
-          throw err; // or handle the error
-        }
-        zip.file(filename, data, { binary: true });
-        count++;
-        if (count === urls.length) {
-          zip.generateAsync({ type: "blob" }).then(function (content) {
-            saveAs(content, zipFilename);
-          });
-        }
-      });
-      return console.log("this needed a return");
-    });
-  };
   // Handles what happens when save plot button is clicked
   const handleClick = () => {
     let datalist = [];
@@ -68,7 +39,7 @@ export default function DownloadDataButton({
       )
     );
     // call zip function here
-    return downloadzip(datalist);
+    return downloadzip(datalist, "data");
   };
 
   return (
